+++ src/components/ContentTab/ArcContent/ArcContent.vue
@@ -19,7 +19,6 @@
         <Tags
           v-if="currentArc"
           v-model="currentArc.tags"
-          :tag-setting="SettingKey.contentTags"
           @tag-added="onTagChange"
           @tag-removed="onTagChange"
           @tag-click="onTagClick"