+++ src/utils/migration/versions/index.ts
@@ -5,6 +5,7 @@ import { MigrationV1_8 } from './MigrationV1_8';
 import { MigrationV1_8_5 } from './MigrationV1_8_5';
 import { MigrationV1_8_6 } from './MigrationV1_8_6';
 import { MigrationV1_9 } from './MigrationV1_9';
+import { MigrationV1_9_1 } from './MigrationV1_9_1';
 
 export const migrationVersions: Record<string, MigrationConstructor> = {
   '1.6.0': MigrationV1_6,
@@ -13,4 +14,5 @@ export const migrationVersions: Record<string, MigrationConstructor> = {
   '1.8.5': MigrationV1_8_5,
   '1.8.6': MigrationV1_8_6,
   '1.9.0': MigrationV1_9,
+  '1.9.1': MigrationV1_9_1,
 };