+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Change Log
+## 1.9.1 - Tag tweaks
+![](https://img.shields.io/badge/release%20date-February%20____%2C%202026-blue)
+![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.9.1/module.zip)
+
+- Fixed issue where the tag list didn't show the autocomplete dropdown with available tags.
+- Tags are now per-Setting instead of global across the entire application.
+- Tag autocomplete list colors are now better in dark mode. 
+
 ## 1.9.0 - Some big quality of life feature requests
 ![](https://img.shields.io/badge/release%20date-February%2027%2C%202026-blue)
 ![GitHub release](https://img.shields.io/github/downloads/dovrosenberg/fvtt-campaign-builder/v1.9.0/module.zip)